+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added method `DatabaseManager.entryCount()`
 - Added method `PluginCommands.entryCount()`
 - Added property `PositionPolling.meta` for access to plugin metadata
+- Added table `plugin_info` with `version` column consisting of a single row
 
 ## [0.2.0] - 2026-08-10